Add a location parameter in resource source
It could be nice to be able to set the location (e.g.: Europe/Paris) for the user like this he could not care about time saving in his coun
try.
For now, we have to set the time zone in start/stop parameter (e.g.: `3 PM +0800`) which are able with this feature just to set the location and omit the time zone, new start and stop values could be simply `3 PM`.

This feature do not change the actual behaviour, the defaults location is set to UTC like it was before.