add 'location' field, used for start/stop/days
previously the timezone offset would be specified in the 'start' and
'stop' fields, e.g. 3:04 PM -0700.

the 'days' field could also be specified, but did not indicate a time
zone, so we'd always compare against UTC, which is unintuitive if
'start' and 'stop' are both specifying a timezone, as it wouldn't be
respected.

this adds a 'location' field, and, if present, it will be used to handle
both. 'start' and 'stop' can just be given as a timestamp without a
timezone, e.g. 3:04 PM or 15:04, and we'll interpret it in the given
location. 'days' will also respect this same field.

as part of this, I pushed all of the 'check' logic down into a package
that can be unit tested, absorbing the 'between' package into it. the
'check' integration tests are now quite redundant but I've left them
there for now.

[finishes #127603783 #127501161]