Fix checking whether previous time is in range
The existing algorithm checks whether the day of the previous time
is less than the day of the current time, when the two times are
converted to UTC.

This is problematic in two ways:
- It does not ensure that the previous is not in range when the current
time is: the two times could in fact be less than 24 hours apart.
- It does not allow the resource to trigger after the range has been moved
forward slightly, and Concourse has cached a previous time
less than a day before the current time, even when the previous time
is out of the new range.

This commit fixes the above problems.

This commit also adds new tests which verify that the problems described
above are fixed.

Moreover, the tests are fixed by interpreting test times in the year
2018 instead of the year 0. Without this some tests would fail falsely,
because they expect present time to be in the range when the range is
interpreted at the given location in 2018, but that is not the case when
it is interpreted at the given location in the year 0.